generate_feature_blob/requirements.txt
insightface==0.7.3
onnxruntime==1.24.2
numpy==2.4.2
opencv-python-headless==4.13.0.92
pymysql==1.1.2
sqlalchemy==2.0.47
python-dotenv==1.2.1
cryptography==46.0.5
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
requests==2.32.5
psycopg2-binary==2.9.11
